requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
google-api-python-client
pandas
nltk==3.6.3
sklearn